refactor; improved error message
