+++ cognite/neat/_issues/errors/_wrapper.py
@@ -5,7 +5,7 @@ from cognite.neat._issues import NeatError
 
 @dataclass(unsafe_hash=True)
 class MetadataValueError(NeatError, ValueError):
-    """{field_name}: {error}"""
+    """Field {field_name} - {error}"""
 
     field_name: str
     error: NeatError
